testHtmIndex: use lower HTM depth for faster test
This test used to take 372 sec because it used an HTM depth of 8 (mean pixel
area of 0.08 deg^2) and a 3 deg search radius. Reducing the HTM depth to 4
(mean pixel area of 20 deg^2, roughly matching the 28 deg^2 search area;
big enough to chunk the data reasonably, small enough to force the code
to check neighbours) allows the test to run in 27 sec.

Changing the HTM depth required regenerating a comparison file; added a hook
to make this easier in the future. The new comparison file has multiple
sources in it, so robustified the test to work regardless of the number of
sources.